split set_gpio_val callback
Especially for controlling the lcd, where we are toggling different gpio lines
hundreds of times in a second. It's better to keep the fd of the GPIO valud
entry around, instead of calling open/close every time when we want to access a
GPIO line. Hence the set_gpio_val callback is splitted into set_gpio_val_by_fd
and gpio_get_val_fd, which can be called independently. Also made change in the
lcd callback to make use of the new callbacks. An array of fds needs to be
created, some addition to close_lcd routine, nothing complicated.