compare song information with stored line
it's unneccessary to fetch name and title from old mpd_song. Actually the
information is stored in the fline and sline buffer. Comparing the new song
information with the line buffers now. This way we can also fill the lines with
other information, while the content will be refreshed automatically.