Added start script
This script can be called while starting up. I used two remaining GPIO as I/O
to watch the on/off trigger and to turn on/off the backlight of the LCD.
(Raspberry Pi has 8 GPIOs as for default, while 6 are used now for the lcd).
This script exports the GPIOs if necessary and than starts to watch the input
GPIO line, while a high flank is detected, it turns on the back light and calls
the radioberry binary. Otherwise it terminates the radioberry process and turn
off the backlight.

Signed-off-by: Luotao Fu <devtty0@gmail.com>